Improve the template
- Rename "closed" to "merged" (closed PRs are ignored already)
- Show author instead of committer (which is bors) for merged PRs
- Show the repo in the PR description
- Hide PR+commit messages by default for merged PRs

The last item is a mildly opinionated stance: since the newly opened PRs
are shown first, they are evidently supposed to be more interesting to
the reader. When a PR merges, the digest affords a second chance to take
a look at it, but presumably the reader already knows the PR from an
earlier digest, so by default it ought to be collapsed.